pyproject.toml
[project]
name = "potrysite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "asyncpg>=0.31.0",
    "dotenv>=0.9.9",
    "fastapi[standart]>=0.128.0",
    "psycopg>=3.3.2",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
